feat: consolida enpoint
