+++ src/main/java/com/fiap/mspedidoapi/domain/useCase/pedido/IniciaPreparoPedidoUseCase.java
@@ -3,7 +3,7 @@ package com.fiap.mspedidoapi.domain.useCase.pedido;
 import com.fiap.mspedidoapi.domain.entity.entrega.Entrega;
 import com.fiap.mspedidoapi.domain.enums.pedido.StatusPedido;
 import com.fiap.mspedidoapi.domain.exception.pedido.PedidoNaoEncontradoException;
-import com.fiap.mspedidoapi.domain.gateway.pedido.BuscaListaPedidoInterface;
+import com.fiap.mspedidoapi.domain.gateway.pedido.PreparaPedidoInterface;
 import com.fiap.mspedidoapi.domain.generic.output.OutputError;
 import com.fiap.mspedidoapi.domain.generic.output.OutputInterface;
 import com.fiap.mspedidoapi.domain.generic.output.OutputStatus;
@@ -18,30 +18,40 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class IniciaPreparoPedidoUseCase {
     
-    private final BuscaListaPedidoInterface buscaListaPedidoInterface;
+    private final PreparaPedidoInterface preparaPedidoInterface;
     private OutputInterface outputInterface;
 
     public void execute(UUID pedidoUUID, Integer tempoDePreparoEmMinutos) {
         try {
-            Pedido pedidoEncontrado = buscaListaPedidoInterface.encontraPedidoPorUuid(pedidoUUID);
-        
-            if(pedidoEncontrado.getStatusPedido() != StatusPedido.RECEBIDO) {
-                this.outputInterface = new PedidoProntoOutput(
-                        null,
-                        new OutputStatus(404, "Not found", "Pedido precisa estar com status RECEBIDO para dar inicio ao preparo")
+
+            if(tempoDePreparoEmMinutos <= 0){
+                this.outputInterface = new OutputError(
+                    "Error", 
+                    new OutputStatus(404, "Falha", "Tempo de preparo precisa ser positivo")
                 );
                 return;
             }
 
-            Entrega entrega = buscaListaPedidoInterface.movePedidoParaEmPreparacao(pedidoUUID, tempoDePreparoEmMinutos);
+            Pedido pedidoEncontrado = preparaPedidoInterface.encontraPedidoPorUuid(pedidoUUID);
+
+            if(pedidoEncontrado.getStatusPedido() != StatusPedido.RECEBIDO) {
+                this.outputInterface = new OutputError(
+                    "Error", 
+                    new OutputStatus(404, "Not found", "Pedido precisa estar com status RECEBIDO para dar inicio ao preparo")
+                );
+                return;
+            }
+
+            Entrega entrega = preparaPedidoInterface.movePedidoParaEmPreparacao(pedidoUUID, tempoDePreparoEmMinutos);
             this.outputInterface = new PedidoProntoOutput(
                 entrega,
                 new OutputStatus(200, "OK", "Pedido atualizado.")
             );
 
         }catch (PedidoNaoEncontradoException e) {
-            this.outputInterface = new PedidoProntoOutput(
-                    null,
+            System.out.println("Pedido nao enconstrado");
+            this.outputInterface = new OutputError(
+                    e.getMessage(),
                     new OutputStatus(404, "Not found", "Pedido não encontrado")
             );
         } catch (Exception e) {
