+++ src/main/java/com/fiap/mspedidoapi/infra/adpter/repository/pedido/BuscarListaPedidoRepository.java
@@ -1,48 +1,18 @@
 package com.fiap.mspedidoapi.infra.adpter.repository.pedido;
 
-import com.fiap.mspedidoapi.domain.entity.entrega.Entrega;
 import com.fiap.mspedidoapi.domain.entity.pedido.PedidoEntity;
 import com.fiap.mspedidoapi.domain.entity.pedido.ProdutoEntity;
-import com.fiap.mspedidoapi.domain.enums.pedido.StatusPedido;
-import com.fiap.mspedidoapi.domain.exception.pedido.PedidoNaoEncontradoException;
 import com.fiap.mspedidoapi.domain.gateway.pedido.BuscaListaPedidoInterface;
-import com.fiap.mspedidoapi.infra.collection.pedido.Pedido;
 import com.fiap.mspedidoapi.infra.collection.pedido.items.Produto;
 import com.fiap.mspedidoapi.infra.repository.PedidosMongoRepository;
 import lombok.RequiredArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 @RequiredArgsConstructor
 public class BuscarListaPedidoRepository implements BuscaListaPedidoInterface {
     public final PedidosMongoRepository pedidosMongoRepository;
-
-    @Override
-    public Pedido encontraPedidoPorUuid(UUID pedidoUuid) throws PedidoNaoEncontradoException {
-        Optional<Pedido> pedidoEncontrado = pedidosMongoRepository.findByUuidPedido(pedidoUuid);
-        if(pedidoEncontrado.isPresent()){
-            return pedidoEncontrado.get();
-        }else{
-            throw new PedidoNaoEncontradoException("Pedido não encontrado");
-        }
-    }
-    
-    @Override
-    public Entrega movePedidoParaEmPreparacao(UUID pedidoUuid, Integer tempoDePreparoEmMinutos) throws PedidoNaoEncontradoException {
-        Optional<Pedido> pedido = pedidosMongoRepository.findByUuidPedido(pedidoUuid);
-        if(pedido.isPresent()){
-            Pedido pedidoEncontrado = pedido.get();
-            pedidoEncontrado.setTempoDePreparo(tempoDePreparoEmMinutos);
-            pedidoEncontrado.setStatusPedido(StatusPedido.EM_PREPARACAO);
-            pedidosMongoRepository.save(pedidoEncontrado);
-            return new Entrega(pedidoEncontrado.getUuidPedido(), pedidoEncontrado.getNumeroPedido(), pedidoEncontrado.getStatusPedido());
-        }else{
-            throw new PedidoNaoEncontradoException("Pedido não encontrado");
-        }
-    }
 
     private static List<ProdutoEntity> getProdutoEntities(List<Produto>  produtosDoPedido) {
         List<ProdutoEntity> produtosList = new ArrayList<>();
