+++ src/test/java/com/fiap/mspedidoapi/infra/adpter/repository/pedido/PedidoProntoRepositoryTest.java
@@ -13,7 +13,10 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 public class PedidoProntoRepositoryTest {
